# Read Parquet files faster