… src/Common/RangeGenerator.h
… src/Common/tests/gtest_range_generator.cpp
… src/Formats/FormatFactory.cpp
… src/Formats/FormatFactory.h
… src/IO/BoundedReadBuffer.cpp
… src/IO/BoundedReadBuffer.h
… src/IO/ParallelReadBuffer.cpp
… src/IO/ParallelReadBuffer.h
… src/IO/ReadBuffer.h
… src/IO/ReadBufferFromFileDescriptor.h
… src/IO/ReadBufferFromS3.cpp
… src/IO/ReadBufferFromS3.h
… src/IO/ReadSettings.h
… src/IO/ReadWriteBufferFromHTTP.h
… src/IO/SeekableReadBuffer.h
… src/Interpreters/Cache/FileCache.h
… src/Interpreters/Cache/FileSegment.cpp
… src/Interpreters/Cache/FileSegment.h
… src/Processors/Formats/IInputFormat.cpp
… src/Processors/Formats/IInputFormat.h
… src/Processors/Formats/IRowInputFormat.cpp
… src/Processors/Formats/ISchemaReader.h
… src/Processors/Formats/Impl/ArrowBlockInputFormat.cpp
… src/Processors/Formats/Impl/ArrowBufferedStreams.cpp
… src/Processors/Formats/Impl/ArrowBufferedStreams.h
… src/Processors/Formats/Impl/JSONColumnsBlockInputFormatBase.cpp
… src/Processors/Formats/Impl/NativeFormat.cpp
… src/Processors/Formats/Impl/ORCBlockInputFormat.cpp
… src/Processors/Formats/Impl/ParallelParsingInputFormat.h
… src/Processors/Formats/Impl/ParquetBlockInputFormat.cpp