@al13n321 2023-03-30
Some speed numbers. Better than before, worse than what is possible.
* Querying one column from a local file didn't get faster (or slower), it's still reading sequentially. (`SELECT AdvEngineID, COUNT(*) AS c FROM file('ds/hits.parquet') GROUP BY AdvEngineID ORDER BY c DESC LIMIT 10` 0.75 s)
* All columns, local file: ~8x faster, it's decoding in parallel (after reading sequentially). (`SELECT sum(ignore(*)) AS c FROM file('ds/hits.parquet')` 102 s -> 12.7 s)
* One column, S3, local region: 50x faster, it's both reading and decoding in parallel. (`SELECT AdvEngineID, COUNT(*) AS c FROM url('https://clickhouse-datasets-us-west-2.s3.amazonaws.com/hits.parquet') GROUP BY AdvEngineID ORDER BY c DESC LIMIT 10` 49 s -> 1 s)
* All columns, S3, local region: 4x faster. (`SELECT sum(ignore(*)) AS c FROM url('https://clickhouse-datasets-us-west-2.s3.amazonaws.com/hits.parquet')` 193 s -> 46 s)
* One column, remote region (US <- Europe): 27x faster (`SELECT AdvEngineID, COUNT(*) AS c FROM url('https://clickhouse-public-datasets.s3.amazonaws.com/hits_compatible/hits.parquet') GROUP BY AdvEngineID ORDER BY c DESC LIMIT 10` 385 s -> 14 s)
* Increasing max_threads from 16 to 128 makes the previous query another 3x faster (5 s). Increasing to 256 (more than the number of HTTP requests the query does) does nothing, still 5x slower than local region, not sure why.
* All columns, remote region: 7x faster (`SELECT sum(ignore(*)) AS c FROM url('https://clickhouse-public-datasets.s3.amazonaws.com/hits_compatible/hits.parquet')` 800 s -> 115 s)
* All of the above gets about the same speed between `url()` and `s3()` table functions. (As it should, S3 is just HTTP with extra steps.) (Before this PR, this wasn't always the case; the comparisons above are using `url()`.)
* If the query does anything nontrivial with the fetched data, it gets bottlenecked on https://github.com/ClickHouse/ClickHouse/issues/38755 instead.