@Avogar 2023-03-30
Looks good in general. I will go deep into details later.
Also, it would be really good to try to refactor/simplify code in `FormatFactory::getInputImpl` and in `ParquetBlockInputFormat`, because it was a bit hard to read it and understand all logic.