@al13n321 2023-04-04
Tests still fail, at least some of them because order of rows is not preserved by default anymore.

Should we make `input_format_parquet_preserve_order` setting default to true, in case anyone relies on the ordering? It's tempting to default to false to get better speed out of the box, especially in benchmarks.