@al13n321 2023-04-04
Would be nice to add a virtual column for row index, then detect `ORDER BY _row_idx` to switch to order-preserving mode. Doesn't help with breaking people's existing queries, just a prettier interface than a setting.